--- simple-chat/src/chat-list.css
@@ -1,126 +1,0 @@
-* {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-}
-
-body{
-    font-family: Arial, sans-serif;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background-color: #e5e5e5;
-}
-
-button {
-    border: 0;
-    background: 0;
-}
-
-.main-container {
-    width: 400px;
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    background-color: #fff;
-    border: 1px solid #ccc;
-    overflow: hidden;
-}
-
-.chat-list-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding-top: 10px;
-    padding-bottom: 10px;
-    background-color: #0088cc;
-    color: #fff;
-    min-height: 56px;
-    gap: 10px;
-}
-
-.burger-button, .search-buttonn {
-    color: #fff;
-    font-size: 24px;
-    cursor: pointer;
-}
-
-.chat-name {
-    font-size: 16px;
-}
-
-.chat-list-container {
-    flex: 1;
-    padding: 15px;
-    overflow-y: scroll; 
-    background-color: #e5ddd5;
-    scrollbar-width: none; 
-}
-
-.message-time {
-    font-size: 12px;
-    color: #888;
-}
-
-.new-chat-button {
-    position: fixed;
-    z-index: 100;
-    bottom: 30px;
-    right: 600px;
-    border: none;
-    color: #0088cc;
-    background-color: #ccc;
-    border-radius: 16px;
-    transform: scale(2.5);
-    cursor: pointer;
-}
-
-.new-chat-icon {
-    font-size: 20px;
-    gap: 2px;
-}
-
-.chat-link {
-    text-decoration: none;
-    color: #202020;
-}
-
-.left-container {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-}
-
-#chat1 {
-    border-top: 1px solid;
-
-}
-
-.chat-item-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    border-bottom: 1px solid;
-}
-
-.chat-status {
-    display: flex;
-    align-items: center;
-    gap: 4px;
-}
-
-.status-time {
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-}
-
-.message-count {
-    background-color: #b3b3b3;
-    border-radius: 4px;
-}
-
-.chat-avatar {
-    font-size: 36px;
-}
