+++ simple-chat/src/chat.css
@@ -1,0 +1,193 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+body{
+    font-family: Arial, sans-serif;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #e5e5e5;
+}
+
+button {
+    border: 0;
+    background: 0;
+}
+
+.main-container {
+    width: 400px;
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+    background-color: #fff;
+    border: 1px solid #ccc;
+    overflow: hidden;
+}
+
+/* Шапка чата */
+.chat-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    background-color: #0088cc;
+    color: #fff;
+    min-height: 56px;
+    gap: 10px;
+}
+
+.back-button, .search-button, .more-button {
+    color: #fff;
+    font-size: 24px;
+    cursor: pointer;
+}
+
+.avatar {
+    font-size: 36px;
+}
+
+.user-container {
+    display: flex;
+    justify-content: center;
+    gap: 6px;
+}
+
+.user-info {
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+}
+
+.chat-name {
+    font-size: 16px;
+}
+
+.last-seen {
+    font-size: 12px;
+    color: #dcdcdc;
+}
+
+/* Контейнер сообщений */
+.chat-container {
+    flex: 1;
+    padding: 15px;
+    overflow-y: scroll; 
+    background-color: #e5ddd5;
+    scrollbar-width: none;
+    scroll-behavior: smooth;  
+}
+
+.messages-container {
+    overflow-y: auto; 
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    gap: 6px;
+}
+
+/* Сообщения */
+.message {
+    padding: 10px;
+    border-radius: 10px;
+    width: auto;
+    max-width: 80%;
+    font-size: 15px;
+    color: #202020;
+}
+
+.message-content {
+    word-wrap: break-word;
+    overflow-wrap: break-word;
+    max-width: 100%;
+    width: auto;
+}
+
+.sender {
+    background-color: #dcf8c6;
+    margin-left: auto;
+}
+
+.receiver {
+    background-color: #ffffff;
+    margin-right: auto;
+    color: #202020;
+}
+
+.message img {
+    max-width: 100%;
+    max-height: 200px;
+    border-radius: 8px;
+}
+
+.message-info {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    gap: 5px;
+}
+
+.message-time {
+    font-size: 12px;
+    color: #888;
+}
+
+.message-check {
+    font-size: 14px;
+    color: #34b7f1;
+}
+
+/* Форма */
+.message-form {
+    display: flex;
+    padding: 10px;
+    border-top: 1px solid #ccc;
+    background-color: #f7f7f7;
+    gap: 10px;
+}
+
+.message-input {
+    flex-grow: 1;
+    padding: 10px;
+    border-radius: 20px;
+    border: 1px solid #ccc;
+    resize: none;
+    outline: none;
+    font-size: 14px;
+    scrollbar-width: none;
+}
+
+.image-input {
+    display: none;
+}
+
+.image-button, .send-button {
+    font-size: 24px;
+    cursor: pointer;
+    color: #0088cc;
+}
+
+.message-content p {
+    white-space: pre-wrap;
+}
+
+.delete-button {
+    position: fixed;
+    z-index: 100;
+    bottom: 30px;
+    right: 30px;
+    border: none;
+    color: #0088cc;
+    background-color: none;
+    border-radius: 16px;
+    transform: scale(2.5);
+    cursor: pointer;
+}
+
+.delete-icon {
+    font-size: 20px;
+}
