+++ simple-chat/src/chat.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+<head lang="en">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Chat App</title>
+    <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+</head>
+    <body>
+        <div class="main-container">
+
+            <div id="chat-header" class="chat-header">
+
+            </div>
+
+            <div id="chat-container" class="chat-container">
+                <div id="messages-container" class="messages-container">     
+
+                </div>   
+            </div>
+        
+            <form id="message-form" class="message-form">
+                <textarea id="message-input" class="message-input" placeholder="Введите сообщение" rows="1"></textarea>
+                <input type="file" id="image-input" class="image-input" accept="image/*"">
+                <button type="button" id="image-button" class="image-button material-icons">image</button>
+                <button type="submit" class="send-button material-icons">send</button>
+            </form>
+
+            <button id="delete-button" class="delete-button">
+                <span class="delete-icon material-icons">cancel</span>
+            </button>
+
+        </div>
+    </body>
+</html>
