+++ simple-chat/src/components/ChatHeader.js
@@ -1,6 +1,6 @@
 export function ChatHeader() {
     return `                
-        <a href="chat-list.html"><span class="back-button material-icons">arrow_back</span></a>
+        <a href="index.html"><span class="back-button material-icons">arrow_back</span></a>
         <div class="user-container">
             <span class="avatar material-icons">account_circle</span> 
             <div class="user-info">
